remove all crud of the status
