+++ www/Views/layout/back.tpl.php
@@ -174,7 +174,6 @@
                     <li class="submenu-item"><a class="submenu-link" href="/back/type">Type annonce</a></li>
                     <li class="submenu-item"><a class="submenu-link" href="/back/user">Liste des utilisateurs</a></li>
                     <li class="submenu-item"><a class="submenu-link" href="/back/agent">Liste des agents</a></li>
-                    <li class="submenu-item"><a class="submenu-link" href="/back/status">Liste des Status</a></li>
                     <li class="submenu-item"><a class="submenu-link" href="/back/opinion-list">Liste des opinions</a></li>
                   </ul>
                 </div>
